requirements.txt
django==1.5.0
south
django-cuser==2014.9.28
